Implements `@service` decorator
Primary implemntation is using Stage 3 API, but provides a Babel
"legacy" decorators shim. TypeScript "experimental" decorators is
supported as types only, trying to call them as such at runtime
will result in an error.

It is only intended to be used on fields and accessors, and if
using TypeScript, it will be enforced. However, it also detects
all forms of invalid usages in debug mode and provide good error
messages. The patterns here should probably be replicated in Ember
and elsewhere.

Continue discussion of the preferred/recommended style in #7.